add loop counter and final update count output
when the job unexpectedly finds no annotated secrets there is no output
at all, making the user question whether it ran or what happened. By
adding a counter and a final statement of the number of items updated
the user can see that no items were updated and know that there was work
attempted but nothing happened

note that this involved moving the code around. I had to keep the last
add loop counter and final update count output

when the job unexpectedly finds no annotated secrets there is no output
at all, making the user question whether it ran or what happened. By
adding a counter and a final statement of the number of items updated
the user can see that no items were updated and know that there was work
attempted but nothing happened

note that this involved moving the code around. I had to keep the last
pipe step in the current shell rather than a subshell. see
https://tclish.org/2019/02/14/the-lastpipe-maneuver/
pipe step in the current shell rather than a subshell.
see https://tclish.org/2019/02/14/the-lastpipe-maneuver/ for examples
and lastpipe did not work in my mac's bash for testing so I went with
the standard, even if it is slightly less readable